two levels, manhattan distance, zooming viewport
the two levels are merged into scene.cpp, there is one int called level that is apart of the scene class. When the player presses v it is set to one, when the player presses z it is set to two. The scene checks the value of level int to decide waht to draw or to erase in order to create the first or second level. Manhattan distance funciton was added, attempt to make the viewport zoom in as the players get closer to one another