Gameplay Changes
when the projectile a collides with the ball it causes the ball to change its x direction only. Reinstituted player gravity it it is working. Projectile A is now deleted when it collides with the Ball. This made it easier to control the directino of the ball when you throw your proctile at it. Ball speed is faster normally and the player's hit makes it speed up at a rate slower than in previous versions. Aim speed is increased by double.